add compute_angle to analysis.py
