+++ simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/data_absz/1water_data/h_density.txt
@@ -99,7 +99,7 @@
 -2.999999999999913847e-02 0.000000000000000000e+00 0.000000000000000000e+00
 -1.999999999999912959e-02 0.000000000000000000e+00 0.000000000000000000e+00
 -9.999999999999120703e-03 0.000000000000000000e+00 0.000000000000000000e+00
-8.881784197001252323e-16 1.383942272480685742e-01 1.165850913667857564e-01
+8.881784197001252323e-16 2.767884544961371485e-01 2.331701827335715127e-01
 1.000000000000089706e-02 3.244989097060171912e-01 2.140986883914658412e-01
 2.000000000000090594e-02 5.018461441871454598e-01 3.279574912193808922e-01
 3.000000000000091482e-02 3.598381895972344036e-01 2.607898982099254859e-01
