+++ simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/data_absz/1water_data/o_density.txt
@@ -99,7 +99,7 @@
 -2.999999999999913847e-02 0.000000000000000000e+00 0.000000000000000000e+00
 -1.999999999999912959e-02 0.000000000000000000e+00 0.000000000000000000e+00
 -9.999999999999120703e-03 0.000000000000000000e+00 0.000000000000000000e+00
-8.881784197001252323e-16 4.487803331768019932e-02 7.413613503872071298e-02
+8.881784197001252323e-16 8.975606663536039864e-02 1.482722700774414260e-01
 1.000000000000089706e-02 1.003203782831019186e-01 1.783451414165730387e-01
 2.000000000000090594e-02 8.755855023961778283e-02 1.124522667115104291e-01
 3.000000000000091482e-02 1.593438329842160328e-01 2.007095897738291335e-01
