+++ simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/data_absz/24water_data/h_density.txt
@@ -99,7 +99,7 @@
 -2.999999999999913847e-02 0.000000000000000000e+00 0.000000000000000000e+00
 -1.999999999999912959e-02 0.000000000000000000e+00 0.000000000000000000e+00
 -9.999999999999120703e-03 0.000000000000000000e+00 0.000000000000000000e+00
-8.881784197001252323e-16 2.891542112610781956e+01 3.124279514938790303e+00
+8.881784197001252323e-16 5.783084225221563912e+01 6.248559029877580606e+00
 1.000000000000089706e-02 5.581883147209274654e+01 5.565739679297719178e+00
 2.000000000000090594e-02 5.097500520290007842e+01 4.396474938011838418e+00
 3.000000000000091482e-02 4.779610442491231481e+01 3.134686562347305028e+00
