+++ simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/data_absz/24water_data/o_density.txt
@@ -99,7 +99,7 @@
 -2.999999999999913847e-02 0.000000000000000000e+00 0.000000000000000000e+00
 -1.999999999999912959e-02 0.000000000000000000e+00 0.000000000000000000e+00
 -9.999999999999120703e-03 0.000000000000000000e+00 0.000000000000000000e+00
-8.881784197001252323e-16 2.348076843773316380e+00 1.617817431585585108e+00
+8.881784197001252323e-16 4.696153687546632760e+00 3.235634863171170217e+00
 1.000000000000089706e-02 4.568826593019282178e+00 2.891367730265296032e+00
 2.000000000000090594e-02 5.706881348248937336e+00 3.063711098145502021e+00
 3.000000000000091482e-02 6.977214545977775906e+00 2.903539165863977978e+00
