--- simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/distribution_data/1water_data/angle.txt
@@ -1,201 +1,0 @@
-# Bins	angle_mean	angle_std
--1.000000000000000000e+00 nan nan
--9.899999999999999911e-01 nan nan
--9.799999999999999822e-01 nan nan
--9.699999999999999734e-01 nan nan
--9.599999999999999645e-01 nan nan
--9.499999999999999556e-01 nan nan
--9.399999999999999467e-01 nan nan
--9.299999999999999378e-01 nan nan
--9.199999999999999289e-01 nan nan
--9.099999999999999201e-01 nan nan
--8.999999999999999112e-01 nan nan
--8.899999999999999023e-01 nan nan
--8.799999999999998934e-01 nan nan
--8.699999999999998845e-01 nan nan
--8.599999999999998757e-01 nan nan
--8.499999999999998668e-01 nan nan
--8.399999999999998579e-01 nan nan
--8.299999999999998490e-01 nan nan
--8.199999999999998401e-01 nan nan
--8.099999999999998312e-01 nan nan
--7.999999999999998224e-01 nan nan
--7.899999999999998135e-01 nan nan
--7.799999999999998046e-01 nan nan
--7.699999999999997957e-01 nan nan
--7.599999999999997868e-01 nan nan
--7.499999999999997780e-01 nan nan
--7.399999999999997691e-01 nan nan
--7.299999999999997602e-01 nan nan
--7.199999999999997513e-01 nan nan
--7.099999999999997424e-01 nan nan
--6.999999999999997335e-01 nan nan
--6.899999999999997247e-01 nan nan
--6.799999999999997158e-01 nan nan
--6.699999999999997069e-01 nan nan
--6.599999999999996980e-01 nan nan
--6.499999999999996891e-01 nan nan
--6.399999999999996803e-01 nan nan
--6.299999999999996714e-01 nan nan
--6.199999999999996625e-01 nan nan
--6.099999999999996536e-01 nan nan
--5.999999999999996447e-01 nan nan
--5.899999999999996358e-01 nan nan
--5.799999999999996270e-01 nan nan
--5.699999999999996181e-01 nan nan
--5.599999999999996092e-01 nan nan
--5.499999999999996003e-01 nan nan
--5.399999999999995914e-01 nan nan
--5.299999999999995826e-01 nan nan
--5.199999999999995737e-01 nan nan
--5.099999999999995648e-01 nan nan
--4.999999999999995559e-01 nan nan
--4.899999999999995470e-01 nan nan
--4.799999999999995381e-01 nan nan
--4.699999999999995293e-01 nan nan
--4.599999999999995204e-01 nan nan
--4.499999999999995115e-01 nan nan
--4.399999999999995026e-01 nan nan
--4.299999999999994937e-01 nan nan
--4.199999999999994849e-01 nan nan
--4.099999999999994760e-01 nan nan
--3.999999999999994671e-01 nan nan
--3.899999999999994582e-01 nan nan
--3.799999999999994493e-01 nan nan
--3.699999999999994404e-01 nan nan
--3.599999999999994316e-01 nan nan
--3.499999999999994227e-01 nan nan
--3.399999999999994138e-01 nan nan
--3.299999999999994049e-01 nan nan
--3.199999999999993960e-01 nan nan
--3.099999999999993872e-01 nan nan
--2.999999999999993783e-01 nan nan
--2.899999999999993694e-01 nan nan
--2.799999999999993605e-01 nan nan
--2.699999999999993516e-01 nan nan
--2.599999999999993427e-01 nan nan
--2.499999999999993339e-01 nan nan
--2.399999999999993250e-01 nan nan
--2.299999999999993161e-01 nan nan
--2.199999999999993072e-01 nan nan
--2.099999999999992983e-01 nan nan
--1.999999999999992895e-01 nan nan
--1.899999999999992806e-01 nan nan
--1.799999999999992717e-01 nan nan
--1.699999999999992628e-01 nan nan
--1.599999999999992539e-01 nan nan
--1.499999999999992450e-01 nan nan
--1.399999999999992362e-01 nan nan
--1.299999999999992273e-01 nan nan
--1.199999999999992184e-01 nan nan
--1.099999999999992095e-01 nan nan
--9.999999999999920064e-02 nan nan
--8.999999999999919176e-02 nan nan
--7.999999999999918288e-02 nan nan
--6.999999999999917399e-02 nan nan
--5.999999999999916511e-02 nan nan
--4.999999999999915623e-02 nan nan
--3.999999999999914735e-02 nan nan
--2.999999999999913847e-02 nan nan
--1.999999999999912959e-02 nan nan
--9.999999999999120703e-03 nan nan
-8.881784197001252323e-16 nan nan
-1.000000000000089706e-02 nan nan
-2.000000000000090594e-02 nan nan
-3.000000000000091482e-02 nan nan
-4.000000000000092371e-02 nan nan
-5.000000000000093259e-02 nan nan
-6.000000000000094147e-02 nan nan
-7.000000000000095035e-02 nan nan
-8.000000000000095923e-02 9.528694152832031250e+01 3.746157455444335938e+01
-9.000000000000096811e-02 6.827838897705078125e+01 1.659611892700195312e+01
-1.000000000000009770e-01 6.893675994873046875e+01 1.938444709777832031e+01
-1.100000000000009859e-01 6.617646026611328125e+01 1.925061225891113281e+01
-1.200000000000009948e-01 5.829509735107421875e+01 7.973082065582275391e+00
-1.300000000000010036e-01 5.828652191162109375e+01 9.649774551391601562e+00
-1.400000000000010125e-01 5.311019134521484375e+01 8.673367500305175781e+00
-1.500000000000010214e-01 5.991389846801757812e+01 1.279555511474609375e+01
-1.600000000000010303e-01 5.579560470581054688e+01 8.827913284301757812e+00
-1.700000000000010392e-01 6.303560256958007812e+01 9.673354148864746094e+00
-1.800000000000010481e-01 6.438328552246093750e+01 1.206776714324951172e+01
-1.900000000000010569e-01 7.272788238525390625e+01 1.006744289398193359e+01
-2.000000000000010658e-01 8.473010253906250000e+01 6.535658359527587891e+00
-2.100000000000010747e-01 9.908445739746093750e+01 3.005856990814208984e+00
-2.200000000000010836e-01 1.153088989257812500e+02 4.974309444427490234e+00
-2.300000000000010925e-01 nan nan
-2.400000000000011013e-01 nan nan
-2.500000000000011102e-01 nan nan
-2.600000000000011191e-01 nan nan
-2.700000000000011280e-01 nan nan
-2.800000000000011369e-01 nan nan
-2.900000000000011458e-01 nan nan
-3.000000000000011546e-01 nan nan
-3.100000000000011635e-01 nan nan
-3.200000000000011724e-01 nan nan
-3.300000000000011813e-01 nan nan
-3.400000000000011902e-01 nan nan
-3.500000000000011990e-01 nan nan
-3.600000000000012079e-01 nan nan
-3.700000000000012168e-01 nan nan
-3.800000000000012257e-01 nan nan
-3.900000000000012346e-01 nan nan
-4.000000000000012434e-01 nan nan
-4.100000000000012523e-01 nan nan
-4.200000000000012612e-01 nan nan
-4.300000000000012701e-01 nan nan
-4.400000000000012790e-01 nan nan
-4.500000000000012879e-01 nan nan
-4.600000000000012967e-01 nan nan
-4.700000000000013056e-01 nan nan
-4.800000000000013145e-01 nan nan
-4.900000000000013234e-01 nan nan
-5.000000000000013323e-01 nan nan
-5.100000000000013411e-01 nan nan
-5.200000000000013500e-01 nan nan
-5.300000000000013589e-01 nan nan
-5.400000000000013678e-01 nan nan
-5.500000000000013767e-01 nan nan
-5.600000000000013856e-01 nan nan
-5.700000000000013944e-01 nan nan
-5.800000000000014033e-01 nan nan
-5.900000000000014122e-01 nan nan
-6.000000000000014211e-01 nan nan
-6.100000000000014300e-01 nan nan
-6.200000000000014388e-01 nan nan
-6.300000000000014477e-01 nan nan
-6.400000000000014566e-01 nan nan
-6.500000000000014655e-01 nan nan
-6.600000000000014744e-01 nan nan
-6.700000000000014833e-01 nan nan
-6.800000000000014921e-01 nan nan
-6.900000000000015010e-01 nan nan
-7.000000000000015099e-01 nan nan
-7.100000000000015188e-01 nan nan
-7.200000000000015277e-01 nan nan
-7.300000000000015365e-01 nan nan
-7.400000000000015454e-01 nan nan
-7.500000000000015543e-01 nan nan
-7.600000000000015632e-01 nan nan
-7.700000000000015721e-01 nan nan
-7.800000000000015810e-01 nan nan
-7.900000000000015898e-01 nan nan
-8.000000000000015987e-01 nan nan
-8.100000000000016076e-01 nan nan
-8.200000000000016165e-01 nan nan
-8.300000000000016254e-01 nan nan
-8.400000000000016342e-01 nan nan
-8.500000000000016431e-01 nan nan
-8.600000000000016520e-01 nan nan
-8.700000000000016609e-01 nan nan
-8.800000000000016698e-01 nan nan
-8.900000000000016787e-01 nan nan
-9.000000000000016875e-01 nan nan
-9.100000000000016964e-01 nan nan
-9.200000000000017053e-01 nan nan
-9.300000000000017142e-01 nan nan
-9.400000000000017231e-01 nan nan
-9.500000000000017319e-01 nan nan
-9.600000000000017408e-01 nan nan
-9.700000000000017497e-01 nan nan
-9.800000000000017586e-01 nan nan
-9.900000000000017675e-01 nan nan
