+++ simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/distribution_data/1water_data/angle_dist.txt
@@ -1,0 +1,101 @@
+# Angle_bins	RelativeFreq
+-9.881796836853027344e-01 7.257753705908676816e-04
+-9.681980609893798828e-01 1.559073123735103773e-03
+-9.482164382934570312e-01 2.123565116811607263e-03
+-9.282348752021789551e-01 3.521353861458980175e-03
+-9.082532525062561035e-01 4.300891375820976165e-03
+-8.882716298103332520e-01 5.322351637930366856e-03
+-8.682900071144104004e-01 5.483637987446336910e-03
+-8.483084440231323242e-01 6.451335318703475818e-03
+-8.283268213272094727e-01 6.720144592458746936e-03
+-8.083451986312866211e-01 6.531973476508460240e-03
+-7.883635759353637695e-01 6.854547484307921788e-03
+-7.683819532394409180e-01 7.956651197471156095e-03
+-7.484003305435180664e-01 8.413611926531472732e-03
+-7.284187078475952148e-01 9.327560694332740973e-03
+-7.084371447563171387e-01 1.048342556423564539e-02
+-6.884555220603942871e-01 8.978113175524885664e-03
+-6.684738993644714355e-01 9.461953348687150711e-03
+-6.484922766685485840e-01 1.053718672097531638e-02
+-6.285107135772705078e-01 1.067157806655908737e-02
+-6.085290908813476562e-01 1.064472055166118883e-02
+-5.885474681854248047e-01 1.072533916513117302e-02
+-5.685658454895019531e-01 1.139735289728224972e-02
+-5.485842227935791016e-01 1.112857148582760587e-02
+-5.286026000976562500e-01 1.193496388300311142e-02
+-5.086210370063781738e-01 1.067157806655908737e-02
+-4.886394143104553223e-01 1.085976541129070674e-02
+-4.686577916145324707e-01 1.034901147512657174e-02
+-4.486761689186096191e-01 1.029525037655448436e-02
+-4.286945760250091553e-01 1.069848176656856831e-02
+-4.087129533290863037e-01 9.945803235835925030e-03
+-3.887313604354858398e-01 9.300670052970892027e-03
+-3.687497377395629883e-01 9.435093216245143619e-03
+-3.487681150436401367e-01 8.816820165822117708e-03
+-3.287865221500396729e-01 9.488833897973194403e-03
+-3.088048994541168213e-01 9.892063542957870922e-03
+-2.888233065605163574e-01 1.048341422155678673e-02
+-2.688416838645935059e-01 1.059095933675380762e-02
+-2.488600611686706543e-01 1.018770613396757611e-02
+-2.288784682750701904e-01 8.978122889532414580e-03
+-2.088968455791473389e-01 8.601794385180758229e-03
+-1.889152228832244873e-01 9.139366980340304925e-03
+-1.689336299896240234e-01 8.924361674625036836e-03
+-1.489520072937011719e-01 9.085645319347175272e-03
+-1.289703994989395142e-01 8.332952246780867683e-03
+-1.089887917041778564e-01 9.327570786430384661e-03
+-8.900717645883560181e-02 8.359868918097548840e-03
+-6.902556866407394409e-02 7.956625371119794773e-03
+-4.904395341873168945e-02 7.768495554116371972e-03
+-2.906234376132488251e-02 7.876017983931130062e-03
+-9.080732241272926331e-03 7.902864388882499883e-03
+1.090087927877902985e-02 7.392167049764713019e-03
+3.088249079883098602e-02 8.359868918097548840e-03
+5.086410045623779297e-02 8.628637649085994787e-03
+7.084571570158004761e-02 6.962077330505675454e-03
+9.082732349634170532e-02 7.311525227403644668e-03
+1.108089387416839600e-01 7.096449655323060907e-03
+1.307905465364456177e-01 5.698688780182251633e-03
+1.507721543312072754e-01 7.069599760320435279e-03
+1.707537770271301270e-01 6.478198359594158386e-03
+1.907353699207305908e-01 5.564285742913802936e-03
+2.107169926166534424e-01 6.101897891987599458e-03
+2.306986153125762939e-01 5.053532330305816021e-03
+2.506802082061767578e-01 5.510524528006422590e-03
+2.706618309020996094e-01 4.919151164025245722e-03
+2.906434535980224609e-01 3.333195324257543727e-03
+3.106250464916229248e-01 3.548224827661530453e-03
+3.306066691875457764e-01 3.628882006248132161e-03
+3.505882620811462402e-01 3.602001398794441988e-03
+3.705698847770690918e-01 3.118136969763163086e-03
+3.905515074729919434e-01 2.876224997544815989e-03
+4.105331003665924072e-01 2.688060745368986513e-03
+4.305147230625152588e-01 2.177319780610484715e-03
+4.504963159561157227e-01 3.118150464628024945e-03
+4.704779386520385742e-01 3.037495496407219449e-03
+4.904595613479614258e-01 2.580549483814919675e-03
+5.104411840438842773e-01 2.499885674034260132e-03
+5.304228067398071289e-01 1.747231922712117348e-03
+5.504043698310852051e-01 1.747247046333018534e-03
+5.703859925270080566e-01 1.290263573695102101e-03
+5.903676152229309082e-01 1.532187993762933664e-03
+6.103492379188537598e-01 1.532201256015108381e-03
+6.303308010101318359e-01 7.526537513221427840e-04
+6.503124237060546875e-01 7.795342424407907716e-04
+6.702940464019775391e-01 6.988988185332073487e-04
+6.902756690979003906e-01 5.644903134916071964e-04
+7.102572917938232422e-01 6.182512957289030631e-04
+7.302389144897460938e-01 5.107337520050361269e-04
+7.502204775810241699e-01 3.763268756610713920e-04
+7.702021002769470215e-01 3.225658934237755253e-04
+7.901837229728698730e-01 2.419265141076487060e-04
+8.101653456687927246e-01 2.688049111864796044e-04
+8.301469683647155762e-01 2.419244200678316169e-04
+8.501285314559936523e-01 2.150457903179099639e-04
+8.701101541519165039e-01 2.419244200678316169e-04
+8.900917768478393555e-01 2.419244200678316169e-04
+9.100733995437622070e-01 1.075228951589549819e-04
+9.300550222396850586e-01 1.344024555932398022e-04
+9.500366449356079102e-01 1.075219644745918282e-04
+9.700182080268859863e-01 2.688072378973874549e-05
+9.899998307228088379e-01 8.064147335594388133e-05
