+++ simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/distribution_data/1water_data/angle_profile.txt
@@ -1,4 +1,4 @@
-# Bins	angle_mean	angle_std
+# Bins	cos_angle_mean	cos_angle_std
 -1.000000000000000000e+00 nan nan
 -9.899999999999999911e-01 nan nan
 -9.799999999999999822e-01 nan nan
@@ -75,54 +75,54 @@
 -2.699999999999993516e-01 nan nan
 -2.599999999999993427e-01 nan nan
 -2.499999999999993339e-01 nan nan
--2.399999999999993250e-01 1.213224716186523438e+02 2.911476135253906250e+00
--2.299999999999993161e-01 1.201030044555664062e+02 4.560820102691650391e+00
--2.199999999999993072e-01 1.145401000976562500e+02 5.040145874023437500e+00
--2.099999999999992983e-01 1.084232788085937500e+02 4.707525730133056641e+00
--1.999999999999992895e-01 1.013386154174804688e+02 4.497457981109619141e+00
--1.899999999999992806e-01 9.556504058837890625e+01 4.737735271453857422e+00
--1.799999999999992717e-01 9.259677124023437500e+01 3.334085464477539062e+00
--1.699999999999992628e-01 9.217388153076171875e+01 3.891089677810668945e+00
--1.599999999999992539e-01 9.120288848876953125e+01 1.779022693634033203e+00
--1.499999999999992450e-01 9.236995697021484375e+01 2.510946989059448242e+00
--1.399999999999992362e-01 9.330164337158203125e+01 2.063686132431030273e+00
--1.299999999999992273e-01 9.338781738281250000e+01 2.705517768859863281e+00
--1.199999999999992184e-01 9.363141632080078125e+01 2.638908147811889648e+00
--1.099999999999992095e-01 9.302392578125000000e+01 3.325584650039672852e+00
--9.999999999999920064e-02 9.301396942138671875e+01 3.973628282546997070e+00
--8.999999999999919176e-02 9.208212280273437500e+01 4.578177452087402344e+00
--7.999999999999918288e-02 8.698428344726562500e+01 5.415260314941406250e+00
--6.999999999999917399e-02 8.342257690429687500e+01 4.701032638549804688e+00
--5.999999999999916511e-02 7.719441986083984375e+01 7.668327331542968750e+00
--4.999999999999915623e-02 7.452060699462890625e+01 1.551546955108642578e+01
--3.999999999999914735e-02 7.490679168701171875e+01 1.937596702575683594e+01
--2.999999999999913847e-02 8.177848815917968750e+01 2.647392463684082031e+01
--1.999999999999912959e-02 8.736508941650390625e+01 2.089359283447265625e+01
--9.999999999999120703e-03 8.956394195556640625e+01 1.682800102233886719e+01
-8.881784197001252323e-16 8.385446929931640625e+01 9.955114364624023438e+00
-1.000000000000089706e-02 8.138265991210937500e+01 1.245892620086669922e+01
-2.000000000000090594e-02 8.181064605712890625e+01 1.229998874664306641e+01
-3.000000000000091482e-02 7.588721466064453125e+01 1.533566284179687500e+01
-4.000000000000092371e-02 7.285385131835937500e+01 1.534151268005371094e+01
-5.000000000000093259e-02 7.475701141357421875e+01 1.026140975952148438e+01
-6.000000000000094147e-02 8.161478424072265625e+01 9.649415969848632812e+00
-7.000000000000095035e-02 8.561071014404296875e+01 8.702408790588378906e+00
-8.000000000000095923e-02 8.958383178710937500e+01 4.123260021209716797e+00
-9.000000000000096811e-02 9.011899566650390625e+01 3.996197700500488281e+00
-1.000000000000009770e-01 9.105670928955078125e+01 3.477587938308715820e+00
-1.100000000000009859e-01 9.399570465087890625e+01 2.940831422805786133e+00
-1.200000000000009948e-01 9.465433502197265625e+01 2.778264522552490234e+00
-1.300000000000010036e-01 9.355104827880859375e+01 2.780189752578735352e+00
-1.400000000000010125e-01 9.381439208984375000e+01 1.436082839965820312e+00
-1.500000000000010214e-01 9.340052032470703125e+01 2.023231744766235352e+00
-1.600000000000010303e-01 9.153888702392578125e+01 3.228142499923706055e+00
-1.700000000000010392e-01 9.079447174072265625e+01 3.568853139877319336e+00
-1.800000000000010481e-01 9.219838714599609375e+01 2.507250070571899414e+00
-1.900000000000010569e-01 9.482616424560546875e+01 1.349992752075195312e+00
-2.000000000000010658e-01 9.982776641845703125e+01 8.404217362403869629e-01
-2.100000000000010747e-01 1.064750671386718750e+02 2.158276796340942383e+00
-2.200000000000010836e-01 1.145114364624023438e+02 3.252035856246948242e+00
-2.300000000000010925e-01 1.186732482910156250e+02 5.463305950164794922e+00
+-2.399999999999993250e-01 nan nan
+-2.299999999999993161e-01 nan nan
+-2.199999999999993072e-01 nan nan
+-2.099999999999992983e-01 nan nan
+-1.999999999999992895e-01 nan nan
+-1.899999999999992806e-01 nan nan
+-1.799999999999992717e-01 nan nan
+-1.699999999999992628e-01 nan nan
+-1.599999999999992539e-01 nan nan
+-1.499999999999992450e-01 nan nan
+-1.399999999999992362e-01 nan nan
+-1.299999999999992273e-01 nan nan
+-1.199999999999992184e-01 nan nan
+-1.099999999999992095e-01 nan nan
+-9.999999999999920064e-02 nan nan
+-8.999999999999919176e-02 nan nan
+-7.999999999999918288e-02 nan nan
+-6.999999999999917399e-02 nan nan
+-5.999999999999916511e-02 nan nan
+-4.999999999999915623e-02 nan nan
+-3.999999999999914735e-02 nan nan
+-2.999999999999913847e-02 nan nan
+-1.999999999999912959e-02 nan nan
+-9.999999999999120703e-03 nan nan
+8.881784197001252323e-16 nan nan
+1.000000000000089706e-02 nan nan
+2.000000000000090594e-02 nan nan
+3.000000000000091482e-02 nan nan
+4.000000000000092371e-02 nan nan
+5.000000000000093259e-02 nan nan
+6.000000000000094147e-02 nan nan
+7.000000000000095035e-02 nan nan
+8.000000000000095923e-02 -7.353657484054565430e-02 5.308204293251037598e-01
+9.000000000000096811e-02 3.202409148216247559e-01 2.205752581357955933e-01
+1.000000000000009770e-01 2.946668565273284912e-01 2.750636339187622070e-01
+1.100000000000009859e-01 3.325950503349304199e-01 2.910300493240356445e-01
+1.200000000000009948e-01 4.590329527854919434e-01 1.193358674645423889e-01
+1.300000000000010036e-01 4.742905199527740479e-01 1.347260922193527222e-01
+1.400000000000010125e-01 5.478315353393554688e-01 1.208277791738510132e-01
+1.500000000000010214e-01 4.375259876251220703e-01 1.905243992805480957e-01
+1.600000000000010303e-01 4.932593703269958496e-01 1.321467906236648560e-01
+1.700000000000010392e-01 3.928243815898895264e-01 1.424482613801956177e-01
+1.800000000000010481e-01 3.710507750511169434e-01 1.713697463274002075e-01
+1.900000000000010569e-01 2.488659620285034180e-01 1.427167952060699463e-01
+2.000000000000010658e-01 6.878077983856201172e-02 1.018533632159233093e-01
+2.100000000000010747e-01 -1.508996486663818359e-01 4.707236588001251221e-02
+2.200000000000010836e-01 -4.015762209892272949e-01 7.055246084928512573e-02
+2.300000000000010925e-01 nan nan
 2.400000000000011013e-01 nan nan
 2.500000000000011102e-01 nan nan
 2.600000000000011191e-01 nan nan
