+++ simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/distribution_data/1water_data/cos_angle_dist.txt
@@ -1,0 +1,101 @@
+# cos_Angle_bins	RelativeFreq
+-9.881796836853027344e-01 3.794416757024664938e-02
+-9.681980609893798828e-01 5.454474088222956152e-02
+-9.482164382934570312e-01 4.031579830433250355e-02
+-9.282348752021789551e-01 4.743020946280831346e-02
+-9.082532525062561035e-01 3.794416757024664938e-02
+-8.882716298103332520e-01 7.351682466735288968e-02
+-8.682900071144104004e-01 5.454474088222956152e-02
+-8.483084440231323242e-01 7.114552641941029876e-02
+-8.283268213272094727e-01 8.774588750619538580e-02
+-8.083451986312866211e-01 7.588833514049329876e-02
+-7.883635759353637695e-01 1.399191179152845244e-01
+-7.683819532394409180e-01 1.351760969690036784e-01
+-7.484003305435180664e-01 1.162040131838803780e-01
+-7.284187078475952148e-01 1.470340879334479656e-01
+-7.084371447563171387e-01 1.422906283884249334e-01
+-6.884555220603942871e-01 1.849778169049524079e-01
+-6.684738993644714355e-01 2.086929216363566097e-01
+-6.484922766685485840e-01 1.897208378512332538e-01
+-6.285107135772705078e-01 2.513808600152497297e-01
+-6.085290908813476562e-01 1.992068797437949179e-01
+-5.885474681854248047e-01 1.636342226466886707e-01
+-5.685658454895019531e-01 1.849778169049524079e-01
+-5.485842227935791016e-01 1.731202645392503625e-01
+-5.286026000976562500e-01 2.347802371840539970e-01
+-5.086210370063781738e-01 2.039499006900757638e-01
+-4.886394143104553223e-01 2.395225577871819833e-01
+-4.686577916145324707e-01 2.324080263677607561e-01
+-4.486761689186096191e-01 2.134362609199593475e-01
+-4.286945760250091553e-01 2.347795368409011374e-01
+-4.087129533290863037e-01 1.992068797437949179e-01
+-3.887313604354858398e-01 3.011822792981648789e-01
+-3.687497377395629883e-01 2.916957881962711219e-01
+-3.487681150436401367e-01 3.367549894514914599e-01
+-3.287865221500396729e-01 3.106678719813944500e-01
+-3.088048994541168213e-01 3.320114662396582150e-01
+-2.888233065605163574e-01 3.509840735128220701e-01
+-2.688416838645935059e-01 3.509835500247815432e-01
+-2.488600611686706543e-01 3.367547383185279797e-01
+-2.288784682750701904e-01 3.841849831521234604e-01
+-2.088968455791473389e-01 4.245003746921344190e-01
+-1.889152228832244873e-01 3.699559097020447829e-01
+-1.689336299896240234e-01 3.486122995269268499e-01
+-1.489520072937011719e-01 3.818134709104437086e-01
+-1.289703994989395142e-01 4.173859989041187046e-01
+-1.089887917041778564e-01 4.197575102615284659e-01
+-8.900717645883560181e-02 4.553303504025166815e-01
+-6.902556866407394409e-02 3.936709587244180963e-01
+-4.904395341873168945e-02 4.980174779043843314e-01
+-2.906234376132488251e-02 4.648163127107586834e-01
+-9.080732241272926331e-03 4.837884071071162917e-01
+1.090087927877902985e-02 4.221291003189543978e-01
+3.088249079883098602e-02 4.245006121184990433e-01
+5.086410045623779297e-02 4.553302655125799348e-01
+7.084571570158004761e-02 4.624448009112140379e-01
+9.082732349634170532e-02 4.600733748858761851e-01
+1.108089387416839600e-01 4.387296011208066115e-01
+1.307905465364456177e-01 4.458441351930359509e-01
+1.507721543312072754e-01 4.743024483359548626e-01
+1.707537770271301270e-01 4.909030340277132920e-01
+1.907353699207305908e-01 5.478193278280278911e-01
+2.107169926166534424e-01 5.193607936177511153e-01
+2.306986153125762939e-01 4.885315217860335402e-01
+2.506802082061767578e-01 5.288472298945896544e-01
+2.706618309020996094e-01 5.288468355103126406e-01
+2.906434535980224609e-01 5.549342783918943534e-01
+3.106250464916229248e-01 5.312183459834531885e-01
+3.306066691875457764e-01 5.715340240268402816e-01
+3.505882620811462402e-01 5.146185402181242541e-01
+3.705698847770690918e-01 5.525619402417168979e-01
+3.905515074729919434e-01 5.976215305758861840e-01
+4.105331003665924072e-01 6.497938696404739201e-01
+4.305147230625152588e-01 6.640229324793164301e-01
+4.504963159561157227e-01 6.592808948416523096e-01
+4.704779386520385742e-01 6.877380372107205764e-01
+4.904595613479614258e-01 6.640229324793164301e-01
+5.104411840438842773e-01 6.189642334896485742e-01
+5.304228067398071289e-01 5.976224219230464874e-01
+5.504043698310852051e-01 6.331932963284909732e-01
+5.703859925270080566e-01 6.569084010598952306e-01
+5.903676152229309082e-01 6.545368905867546827e-01
+6.103492379188537598e-01 7.043386105227035232e-01
+6.303308010101318359e-01 7.754862379715723453e-01
+6.503124237060546875e-01 7.991990294483201085e-01
+6.702940464019775391e-01 8.537437703305496006e-01
+6.902756690979003906e-01 8.063135608677414190e-01
+7.102572917938232422e-01 8.252856446528646917e-01
+7.302389144897460938e-01 9.414924662835295610e-01
+7.502204775810241699e-01 1.003148930138395789e+00
+7.702021002769470215e-01 1.012634972030957536e+00
+7.901837229728698730e-01 1.100380859537152922e+00
+8.101653456687927246e-01 1.015006482504097862e+00
+8.301469683647155762e-01 1.107495390956574122e+00
+8.501285314559936523e-01 1.176272703467583725e+00
+8.701101541519165039e-01 1.079037265278889324e+00
+8.900917768478393555e-01 1.211841851774752454e+00
+9.100733995437622070e-01 1.316188312592930787e+00
+9.300550222396850586e-01 1.297216228807807292e+00
+9.500366449356079102e-01 1.247418229886994023e+00
+9.700182080268859863e-01 1.503537639971023632e+00
+9.899998307228088379e-01 1.745431708231345969e+00
