+++ simulations/nvt-pore/1x1x1.0nm_1-layer/cp2k/distribution_data/1water_data/result.txt
@@ -1,0 +1,4 @@
+The mean of the cos(angle) is 0.379962682723999
+The stdev of the cos(angle) is 0.4918923079967499
+The mean of the angle is 63.77849197387695
+The stdev of the angle is 34.167205810546875
